Contextualize the list of blogs provided to XML-RPC
If the end user takes the address of a section and enters it in a blogging tool, then the call to .getUsersBlogs will either provide a list of sub-sections if there are some, or the target section itself. Previously, it always listed freshest sections as seen in user profile. With the new code it's possible to pick up any section from the content tree and to configure it as the target blog in blogging tool. tested with echo (Mac) and WLW (Windows).